refactor: crate a separation between validator and signer libraries and their executables
