+++ validator/attest_test.go
@@ -1,4 +1,4 @@
-package main_test
+package validator_test
 
 import (
 	"context"
@@ -8,7 +8,7 @@ import (
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/juno/utils"
 	"github.com/NethermindEth/starknet-staking-v2/mocks"
-	main "github.com/NethermindEth/starknet-staking-v2/validator"
+	"github.com/NethermindEth/starknet-staking-v2/validator"
 	"github.com/NethermindEth/starknet.go/rpc"
 	snGoUtils "github.com/NethermindEth/starknet.go/utils"
 	"github.com/sourcegraph/conc"
@@ -24,14 +24,14 @@ func TestProcessBlockHeaders(t *testing.T) {
 	logger := utils.NewNopZapLogger()
 
 	t.Run("Simple scenario: 1 epoch", func(t *testing.T) {
-		dispatcher := main.NewEventDispatcher[*mocks.MockAccounter, *utils.ZapLogger]()
+		dispatcher := validator.NewEventDispatcher[*mocks.MockAccounter, *utils.ZapLogger]()
 		headersFeed := make(chan *rpc.BlockHeader)
 
 		epochId := uint64(1516)
 		epochLength := uint64(40)
 		attestWindow := uint64(16)
-		epochStartingBlock := main.BlockNumber(639270)
-		expectedTargetBlock := main.BlockNumber(639291)
+		epochStartingBlock := validator.BlockNumber(639270)
+		expectedTargetBlock := validator.BlockNumber(639291)
 		mockFetchedEpochAndAttestInfo(
 			t,
 			mockAccount,
@@ -41,7 +41,7 @@ func TestProcessBlockHeaders(t *testing.T) {
 			epochStartingBlock,
 		)
 
-		targetBlockHash := main.BlockHash(*utils.HexToFelt(t, "0x6d8dc0a8bdf98854b6bc146cb7cab6cddda85619c6ae2948ee65da25815e045"))
+		targetBlockHash := validator.BlockHash(*utils.HexToFelt(t, "0x6d8dc0a8bdf98854b6bc146cb7cab6cddda85619c6ae2948ee65da25815e045"))
 		blockHeaders := mockHeaderFeedWithLogger(t, epochStartingBlock, expectedTargetBlock, &targetBlockHash, epochLength)
 
 		// Mock SetTargetBlockHashIfExists call
@@ -59,12 +59,12 @@ func TestProcessBlockHeaders(t *testing.T) {
 		})
 
 		// Events receiver routine
-		receivedAttestEvents := make(map[main.AttestRequired]uint)
+		receivedAttestEvents := make(map[validator.AttestRequired]uint)
 		receivedEndOfWindowEvents := uint8(0)
 		wgDispatcher := conc.NewWaitGroup()
 		wgDispatcher.Go(func() { registerReceivedEvents(t, &dispatcher, receivedAttestEvents, &receivedEndOfWindowEvents) })
 
-		main.ProcessBlockHeaders(headersFeed, mockAccount, logger, &dispatcher)
+		validator.ProcessBlockHeaders(headersFeed, mockAccount, logger, &dispatcher)
 
 		// No need to wait for wgFeed routine as it'll be the 1st closed, causing ProcessBlockHeaders to have returned
 		// Still calling it just in case.
@@ -77,38 +77,38 @@ func TestProcessBlockHeaders(t *testing.T) {
 		// Assert
 		require.Equal(t, 1, len(receivedAttestEvents))
 
-		actualCount, exists := receivedAttestEvents[main.AttestRequired{BlockHash: targetBlockHash}]
+		actualCount, exists := receivedAttestEvents[validator.AttestRequired{BlockHash: targetBlockHash}]
 		require.True(t, exists)
-		require.Equal(t, uint(16-main.MIN_ATTESTATION_WINDOW+1), actualCount)
+		require.Equal(t, uint(16-validator.MIN_ATTESTATION_WINDOW+1), actualCount)
 
 		require.Equal(t, uint8(1), receivedEndOfWindowEvents)
 	})
 
 	t.Run("Scenario: transition between 2 epochs", func(t *testing.T) {
-		dispatcher := main.NewEventDispatcher[*mocks.MockAccounter, *utils.ZapLogger]()
+		dispatcher := validator.NewEventDispatcher[*mocks.MockAccounter, *utils.ZapLogger]()
 		headersFeed := make(chan *rpc.BlockHeader)
 
 		epochLength := uint64(40)
 		attestWindow := uint64(16)
 
 		epochId1 := uint64(1516)
-		epochStartingBlock1 := main.BlockNumber(639270)
-		expectedTargetBlock1 := main.BlockNumber(639291)
+		epochStartingBlock1 := validator.BlockNumber(639270)
+		expectedTargetBlock1 := validator.BlockNumber(639291)
 		mockFetchedEpochAndAttestInfo(t, mockAccount, epochId1, epochLength, attestWindow, epochStartingBlock1)
 
 		epochId2 := uint64(1517)
-		epochStartingBlock2 := main.BlockNumber(639310)
-		expectedTargetBlock2 := main.BlockNumber(639316)
+		epochStartingBlock2 := validator.BlockNumber(639310)
+		expectedTargetBlock2 := validator.BlockNumber(639316)
 		mockFetchedEpochAndAttestInfo(t, mockAccount, epochId2, epochLength, attestWindow, epochStartingBlock2)
 
-		targetBlockHashEpoch1 := main.BlockHash(
+		targetBlockHashEpoch1 := validator.BlockHash(
 			*utils.HexToFelt(t, "0x6d8dc0a8bdf98854b6bc146cb7cab6cddda85619c6ae2948ee65da25815e045"),
 		)
 		blockHeaders1 := mockHeaderFeedWithLogger(
 			t, epochStartingBlock1, expectedTargetBlock1, &targetBlockHashEpoch1, epochLength,
 		)
 
-		targetBlockHashEpoch2 := main.BlockHash(
+		targetBlockHashEpoch2 := validator.BlockHash(
 			*utils.HexToFelt(t, "0x2124ae375432a16ef644f539c3b148f63c706067bf576088f32033fe59c345e"),
 		)
 		blockHeaders2 := mockHeaderFeedWithLogger(
@@ -135,12 +135,12 @@ func TestProcessBlockHeaders(t *testing.T) {
 		})
 
 		// Events receiver routine
-		receivedAttestEvents := make(map[main.AttestRequired]uint)
+		receivedAttestEvents := make(map[validator.AttestRequired]uint)
 		receivedEndOfWindowEvents := uint8(0)
 		wgDispatcher := conc.NewWaitGroup()
 		wgDispatcher.Go(func() { registerReceivedEvents(t, &dispatcher, receivedAttestEvents, &receivedEndOfWindowEvents) })
 
-		main.ProcessBlockHeaders(headersFeed, mockAccount, logger, &dispatcher)
+		validator.ProcessBlockHeaders(headersFeed, mockAccount, logger, &dispatcher)
 
 		// No need to wait for wgFeed routine as it'll be the 1st closed, causing ProcessBlockHeaders to have returned
 		// Still calling it just in case.
@@ -153,13 +153,13 @@ func TestProcessBlockHeaders(t *testing.T) {
 		// Assert
 		require.Equal(t, 2, len(receivedAttestEvents))
 
-		countEpoch1, exists := receivedAttestEvents[main.AttestRequired{BlockHash: targetBlockHashEpoch1}]
+		countEpoch1, exists := receivedAttestEvents[validator.AttestRequired{BlockHash: targetBlockHashEpoch1}]
 		require.True(t, exists)
-		require.Equal(t, uint(16-main.MIN_ATTESTATION_WINDOW+1), countEpoch1)
+		require.Equal(t, uint(16-validator.MIN_ATTESTATION_WINDOW+1), countEpoch1)
 
-		countEpoch2, exists := receivedAttestEvents[main.AttestRequired{BlockHash: targetBlockHashEpoch2}]
+		countEpoch2, exists := receivedAttestEvents[validator.AttestRequired{BlockHash: targetBlockHashEpoch2}]
 		require.True(t, exists)
-		require.Equal(t, uint(16-main.MIN_ATTESTATION_WINDOW+1), countEpoch2)
+		require.Equal(t, uint(16-validator.MIN_ATTESTATION_WINDOW+1), countEpoch2)
 
 		require.Equal(t, uint8(2), receivedEndOfWindowEvents)
 	})
@@ -180,10 +180,10 @@ func sendHeaders(t *testing.T, headersFeed chan *rpc.BlockHeader, blockHeaders [
 
 // Test helper function to register received events to assert on them
 // Note: to exit this function, close the AttestRequired channel
-func registerReceivedEvents[T main.Accounter, Log main.Logger](
+func registerReceivedEvents[T validator.Accounter, Log validator.Logger](
 	t *testing.T,
-	dispatcher *main.EventDispatcher[T, Log],
-	receivedAttestRequired map[main.AttestRequired]uint,
+	dispatcher *validator.EventDispatcher[T, Log],
+	receivedAttestRequired map[validator.AttestRequired]uint,
 	receivedEndOfWindowCount *uint8,
 ) {
 	t.Helper()
@@ -210,7 +210,7 @@ func mockFetchedEpochAndAttestInfo(
 	epochId,
 	epochLength,
 	attestWindow uint64,
-	epochStartingBlock main.BlockNumber,
+	epochStartingBlock validator.BlockNumber,
 ) {
 	t.Helper()
 
@@ -222,7 +222,7 @@ func mockFetchedEpochAndAttestInfo(
 	stake := uint64(1000000000000000000)
 
 	expectedEpochInfoFnCall := rpc.FunctionCall{
-		ContractAddress: utils.HexToFelt(t, main.STAKING_CONTRACT_ADDRESS),
+		ContractAddress: utils.HexToFelt(t, validator.STAKING_CONTRACT_ADDRESS),
 		EntryPointSelector: snGoUtils.GetSelectorFromNameFelt(
 			"get_attestation_info_by_operational_address",
 		),
@@ -245,7 +245,7 @@ func mockFetchedEpochAndAttestInfo(
 
 	// Mock fetchAttestWindow call
 	expectedWindowFnCall := rpc.FunctionCall{
-		ContractAddress:    utils.HexToFelt(t, main.ATTEST_CONTRACT_ADDRESS),
+		ContractAddress:    utils.HexToFelt(t, validator.ATTEST_CONTRACT_ADDRESS),
 		EntryPointSelector: snGoUtils.GetSelectorFromNameFelt("attestation_window"),
 		Calldata:           []*felt.Felt{},
 	}
@@ -262,8 +262,8 @@ func mockFetchedEpochAndAttestInfo(
 func mockHeaderFeedWithLogger(
 	t *testing.T,
 	startingBlock,
-	targetBlock main.BlockNumber,
-	targetBlockHash *main.BlockHash,
+	targetBlock validator.BlockNumber,
+	targetBlockHash *validator.BlockHash,
 	epochLength uint64,
 ) []rpc.BlockHeader {
 	t.Helper()
@@ -272,7 +272,7 @@ func mockHeaderFeedWithLogger(
 
 	blockHeaders := make([]rpc.BlockHeader, epochLength)
 	for i := uint64(0); i < epochLength; i++ {
-		blockNumber := main.BlockNumber(i) + startingBlock
+		blockNumber := validator.BlockNumber(i) + startingBlock
 
 		// All block hashes are set to 0x1 except for the target block
 		if blockNumber == targetBlock {
@@ -301,12 +301,12 @@ func TestSetTargetBlockHashIfExists(t *testing.T) {
 			BlockWithTxHashes(context.Background(), rpc.BlockID{Number: &targetBlockNumber}).
 			Return(nil, errors.New("Block not found"))
 
-		attestInfo := main.AttestInfo{
-			TargetBlock: main.BlockNumber(targetBlockNumber),
+		attestInfo := validator.AttestInfo{
+			TargetBlock: validator.BlockNumber(targetBlockNumber),
 		}
-		main.SetTargetBlockHashIfExists(mockAccount, mockLogger, &attestInfo)
+		validator.SetTargetBlockHashIfExists(mockAccount, mockLogger, &attestInfo)
 
-		require.Equal(t, main.BlockHash{}, attestInfo.TargetBlockHash)
+		require.Equal(t, validator.BlockHash{}, attestInfo.TargetBlockHash)
 	})
 
 	t.Run("Target block already exists but is pending", func(t *testing.T) {
@@ -316,12 +316,12 @@ func TestSetTargetBlockHashIfExists(t *testing.T) {
 			BlockWithTxHashes(context.Background(), rpc.BlockID{Number: &targetBlockNumber}).
 			Return(&rpc.PendingBlockTxHashes{}, nil)
 
-		attestInfo := main.AttestInfo{
-			TargetBlock: main.BlockNumber(targetBlockNumber),
+		attestInfo := validator.AttestInfo{
+			TargetBlock: validator.BlockNumber(targetBlockNumber),
 		}
-		main.SetTargetBlockHashIfExists(mockAccount, mockLogger, &attestInfo)
+		validator.SetTargetBlockHashIfExists(mockAccount, mockLogger, &attestInfo)
 
-		require.Equal(t, main.BlockHash{}, attestInfo.TargetBlockHash)
+		require.Equal(t, validator.BlockHash{}, attestInfo.TargetBlockHash)
 	})
 
 	t.Run("Target block already exists and is not pending", func(t *testing.T) {
@@ -337,17 +337,17 @@ func TestSetTargetBlockHashIfExists(t *testing.T) {
 			BlockWithTxHashes(context.Background(), rpc.BlockID{Number: &targetBlockNumber}).
 			Return(&blockWithTxs, nil)
 
-		targetBlockHash := main.BlockHash(*targetBlockHashFelt)
+		targetBlockHash := validator.BlockHash(*targetBlockHashFelt)
 		mockLogger.EXPECT().
 			Infow(
 				"Target block already exists, registered block hash to attest to it if still within attestation window",
 				"block hash", targetBlockHash.String(),
 			)
 
-		attestInfo := main.AttestInfo{
-			TargetBlock: main.BlockNumber(targetBlockNumber),
+		attestInfo := validator.AttestInfo{
+			TargetBlock: validator.BlockNumber(targetBlockNumber),
 		}
-		main.SetTargetBlockHashIfExists(mockAccount, mockLogger, &attestInfo)
+		validator.SetTargetBlockHashIfExists(mockAccount, mockLogger, &attestInfo)
 
 		require.Equal(t, targetBlockHash, attestInfo.TargetBlockHash)
 	})
