+++ validator/config.go
@@ -1,17 +1,13 @@
-package main
+package validator
 
 import (
-	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"math/big"
 	"os"
 
 	"github.com/NethermindEth/juno/core/crypto"
 	"github.com/NethermindEth/juno/core/felt"
-	"github.com/NethermindEth/juno/utils"
-	"github.com/spf13/cobra"
 )
 
 func isZero[T comparable](v T) bool {
@@ -117,91 +113,6 @@ func checkSigner(signer *Signer) error {
 		return errors.New("neither private key nor url properties set in signer config")
 	}
 	return nil
-}
-
-func NewCommand() cobra.Command {
-	var configPath string
-	var logLevelF string
-
-	var config Config
-	var logger utils.ZapLogger
-
-	preRunE := func(cmd *cobra.Command, args []string) error {
-		if configPath != "" {
-			fileConfig, err := ConfigFromFile(configPath)
-			if err != nil {
-				return err
-			}
-			config.Fill(&fileConfig)
-		}
-		if err := config.Check(); err != nil {
-			return err
-		}
-
-		var logLevel utils.LogLevel
-		if err := logLevel.Set(logLevelF); err != nil {
-			return err
-		}
-		loadedLogger, err := utils.NewZapLogger(logLevel, true)
-		if err != nil {
-			return err
-		}
-		logger = *loadedLogger
-
-		return nil
-	}
-
-	run := func(cmd *cobra.Command, args []string) {
-		if err := Attest(&config, logger); err != nil {
-			logger.Error(err)
-		}
-	}
-
-	var rootCmd = cobra.Command{
-		Use:     "validator",
-		Short:   "Program for Starknet validators to attest to epochs with respect to Staking v2",
-		PreRunE: preRunE,
-		Run:     run,
-		Args:    cobra.NoArgs,
-	}
-
-	// Config file path flag
-	rootCmd.Flags().StringVarP(&configPath, "config", "c", "", "Path to JSON config file")
-
-	// Config provider flags
-	rootCmd.Flags().StringVar(&config.Provider.Http, "provider-http", "", "Provider http address")
-	rootCmd.Flags().StringVar(&config.Provider.Ws, "provider-ws", "", "Provider ws address")
-	// Config signer flags
-	rootCmd.Flags().StringVar(
-		&config.Signer.ExternalUrl,
-		"signer-url",
-		"",
-		"Signer url address, required if using an external signer",
-	)
-	rootCmd.Flags().StringVar(
-		&config.Signer.PrivKey, "signer-priv-key", "", "Signer private key, required for signing",
-	)
-	rootCmd.Flags().StringVar(
-		&config.Signer.OperationalAddress,
-		"signer-op-address",
-		"",
-		"Signer operational address, required for attesting",
-	)
-
-	// Other flags
-	rootCmd.Flags().StringVar(
-		&logLevelF, "log-level", utils.INFO.String(), "Options: trace, debug, info, warn, error.",
-	)
-
-	return rootCmd
-}
-
-func main() {
-	command := NewCommand()
-	if err := command.ExecuteContext(context.Background()); err != nil {
-		fmt.Println("Unexpected error:\n", err)
-		os.Exit(1)
-	}
 }
 
 func ComputeBlockNumberToAttestTo[Account Accounter](
