+++ validator/types.go
@@ -1,4 +1,4 @@
-package main
+package validator
 
 import (
 	"log"
